♻️ Move Option, Options and NewOption to opts
