+++ opts/options_test.go
@@ -1,4 +1,4 @@
-package sqlslog
+package opts
 
 import "testing"
